feat: 🎸 유저 목록 불러오기 스피너 적용 및 더미데이터 제거
✅ Closes: #98
