+++ client/src/components/UserList/index.tsx
@@ -5,7 +5,7 @@ import UserTable from "./UserTable";
 import { useRecoilValue } from "recoil";
 import { socketState } from "../../store/socket";
 import * as style from "./styles";
-import * as dummy from "../../pages/dummy";
+import Spinner from "../Loader/Spinner";
 
 export type User = {
   userId: number;
@@ -17,23 +17,27 @@ export type User = {
 };
 
 export interface UserListProps {
-  userMap: {
-    users: User[];
-    friends: User[];
-    rank: User[];
-  };
+  users: User[];
+  friends: User[];
+  rank: User[];
 }
 
 const UserList = () => {
   const socket = useRecoilValue(socketState);
+  const [loading, setLoading] = useState(true);
   const [selected, setSelected] = useState<number>(0);
-  const [userMap, setUserMap] = useState(dummy.dummyUserMap);
+  const [userMap, setUserMap] = useState<UserListProps>({
+    users: [],
+    friends: [],
+    rank: [],
+  });
 
   useEffect(() => {
     socket.on("user/online", data => {
       setUserMap(current => {
         return { ...current, users: Object.values(data) };
       });
+      setLoading(false);
     });
     return () => {
       socket.off("user/online");
@@ -43,9 +47,18 @@ const UserList = () => {
   return (
     <div css={style.ListContainerStyle}>
       <UserListTab selected={selected} setSelected={setSelected} />
-      {selected === 0 && <UserTable users={userMap.users} selected={selected} />}
-      {selected === 1 && <UserTable users={userMap.friends} selected={selected} />}
-      {selected === 2 && <UserTable users={userMap.rank} selected={selected} />}
+
+      <div css={style.TableContainerStyle}>
+        {loading ? (
+          <Spinner color="#bbbbbb" />
+        ) : (
+          <>
+            {selected === 0 && <UserTable users={userMap.users} selected={selected} />}
+            {selected === 1 && <UserTable users={userMap.friends} selected={selected} />}
+            {selected === 2 && <UserTable users={userMap.rank} selected={selected} />}
+          </>
+        )}
+      </div>
     </div>
   );
 };
